Create featureplotter.py
Function that takes an h5 file containing raw data as an input. For each feature of the input data, will plot a histogram 
of the feature in the ttbar vs. tW data sets and save them as pdfs. Returns a list of features that have statistically significant 
(in this case <.05) p-value outputs of the Kolmogorov-Smirnov statistic to be used when creating x_train and x_test sets from raw data
to be fed into a neural network.